photo_api/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
colorthief==0.2.1
webcolors==1.12
django-cleanup==6.0.0